refactor to use nonzos resource provider
